use 'categories' association for categories
Since we are using a `has_many :through` association for products ->
categories we need to tell ActiveAdmin to use end associtation. For this
to work we need to also allow ActiveAdmin to receive the `category_ids`
paramter. Notice how this is the singlar version of the resource
appended by `_ids`.